+++ mycodo/mycodo_flask/forms/forms_dashboard.py
@@ -85,7 +85,7 @@ class DashboardGraph(FlaskForm):
     enable_title = BooleanField(lazy_gettext('Enable Title'))
     enable_navbar = BooleanField(lazy_gettext('Enable Navbar'))
     enable_export = BooleanField(lazy_gettext('Enable Export'))
-    enable_range = BooleanField(lazy_gettext('Enable Range Selector'))
+    enable_rangeselect = BooleanField(lazy_gettext('Enable Range Selector'))
     enable_graph_shift = BooleanField(lazy_gettext('Enable Graph Shift'))
     enable_manual_y_axis = BooleanField(lazy_gettext('Enable Manual Y-Axis Min/Max'))
     enable_align_ticks = BooleanField(lazy_gettext('Enable Y-Axis Align Ticks'))